+++ src/main.js
@@ -1,3 +1,5 @@
+export * from './motion-input.component';
 export * from './smooth-locomotion.component';
 export * from './smooth-turn.component';
 export * from './snap-turn.component';
+export * from './vignette.primitive';