+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  * Smooth turning component `smooth-turn`
  * Input mode property for switching between analog and binary input handling (`smooth-turn` and `smooth-locomotion`)
  * New RC car example
+ * Vignette primitive `al-vignette` and corresponding example
 
 # 0.1.0 (2022-05-03)
  * Smooth locomotion component `smooth-locomotion`